Modified management of progress action completion status (Issue #42)
The change in the logic for the progress action of an InCommand is driven by
the update of PUS. In the new PUS (April 2016), the progress action logic is
as follows.

During execution, the progress steps are executed in sequence and, if the step's
execution is successful, a (1,5) is generated; if, instead, the step's execution
has failed, a (1,6) is generated (assuming that the acknowledgement of
successful progress was enabled for the command). At the end of the execution of
all the steps, a (1,8) is generated if the command's post-conditions are not
satisfied or if any of its progress steps have failed. Otherwise, a (1,7) may be
generated if it was enabled for that command. Note that this means that
generation of a (1,8) does not necessarily mean that a command's execution has
completely failed. It may simply mean that some of its progress steps have
failed.